# api-spec.yaml
openapi: 3.0.0
info:
  title: 薬飲み忘れ管理API
  description: 薬を飲んだかどうかを管理するWebアプリのAPI
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server

paths:
  /token:
    post:
      summary: ユーザーログイン
      description: ユーザー名とパスワードでログインしてアクセストークンを取得
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: ユーザー名
                password:
                  type: string
                  description: パスワード
              required:
                - username
                - password
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: 認証失敗

  /register:
    post:
      summary: ユーザー登録
      description: 新しいユーザーを登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: ユーザー名が既に存在

  /users/me:
    get:
      summary: 現在のユーザー情報取得
      description: ログイン中のユーザーの情報を取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ユーザー情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: 認証が必要

  /users:
    get:
      summary: 全ユーザー情報取得（管理者限定）
      description: 全ユーザーの情報を取得（管理者のみ）
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ユーザーリスト
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '403':
          description: 管理者権限が必要

  /medication-records:
    get:
      summary: 薬の記録取得
      description: ログイン中のユーザーの薬の記録を取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 薬の記録リスト
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedicationRecordResponse'
    
    post:
      summary: 薬の記録作成
      description: 新しい薬の記録を作成
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationRecordCreate'
      responses:
        '200':
          description: 作成された記録
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationRecordResponse'

  /medication-records/{record_id}:
    put:
      summary: 薬の記録更新
      description: 既存の薬の記録を更新
      security:
        - bearerAuth: []
      parameters:
        - name: record_id
          in: path
          required: true
          schema:
            type: integer
          description: 更新する記録のID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationRecordCreate'
      responses:
        '200':
          description: 更新された記録
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationRecordResponse'
        '404':
          description: 記録が見つからない

  /admin/medication-records:
    get:
      summary: 全ユーザーの薬の記録取得（管理者限定）
      description: 全ユーザーの薬の記録を取得（管理者のみ）
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 全ユーザーの薬の記録リスト
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedicationRecordResponse'
        '403':
          description: 管理者権限が必要

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Token:
      type: object
      properties:
        access_token:
          type: string
          description: JWTアクセストークン
        token_type:
          type: string
          description: トークンタイプ（bearer）
      required:
        - access_token
        - token_type

    UserCreate:
      type: object
      properties:
        username:
          type: string
          description: ユーザー名
        password:
          type: string
          description: パスワード
        is_admin:
          type: boolean
          description: 管理者フラグ
          default: false
      required:
        - username
        - password

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: ユーザーID
        username:
          type: string
          description: ユーザー名
        is_admin:
          type: boolean
          description: 管理者フラグ
        created_at:
          type: string
          format: date-time
          description: 作成日時
      required:
        - id
        - username
        - is_admin
        - created_at

    MedicationRecordCreate:
      type: object
      properties:
        date:
          type: string
          format: date
          description: 記録日付
        morning_taken:
          type: boolean
          description: 朝の服薬フラグ
          default: false
        afternoon_taken:
          type: boolean
          description: 昼の服薬フラグ
          default: false
        evening_taken:
          type: boolean
          description: 夜の服薬フラグ
          default: false
      required:
        - date

    MedicationRecordResponse:
      type: object
      properties:
        id:
          type: integer
          description: 記録ID
        user_id:
          type: integer
          description: ユーザーID
        date:
          type: string
          format: date
          description: 記録日付
        morning_taken:
          type: boolean
          description: 朝の服薬フラグ
        afternoon_taken:
          type: boolean
          description: 昼の服薬フラグ
        evening_taken:
          type: boolean
          description: 夜の服薬フラグ
        created_at:
          type: string
          format: date-time
          description: 作成日時
      required:
        - id
        - user_id
        - date
        - morning_taken
        - afternoon_taken
        - evening_taken
        - created_at